Updating with features around the world

+++ drivers/thermal/supply_lm_core.c
@@ -490,7 +490,7 @@ static void devmgr_mitigation_callback(struct device_clnt_data *clnt,
 					union device_request *req,
 					void *data)
 {
-	uint32_t step1_rel = (int)data;
+	uint32_t step1_rel = (long)data;
 	int ret = 0;
 
 	if (!clnt) {
